requirements.txt
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.3
cffi==1.11.0
decorator==4.1.2
Django==1.11.5
django-crispy-forms==1.6.1
django-extensions==1.9.1
enum34==1.1.6
ipython==5.5.0
ipython-genutils==0.2.0
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pycparser==2.18
Pygments==2.2.0
pytz==2017.2
scandir==1.5
selenium==3.5.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
